--- a/cupy/core/core.pyx
+++ b/cupy/core/core.pyx
@@ -8,6 +8,7 @@ import six
 
 import cupy
 from cupy.core import flags
+from cupy import cuda
 from cupy.cuda import device
 from cupy.cuda import stream
 try:
@@ -26,6 +27,9 @@ from cupy.cuda cimport function
 from cupy.cuda cimport pinned_memory
 from cupy.cuda cimport runtime
 from cupy.cuda cimport memory
+from cupy.core import compile
+from cupy.core import emit
+
 
 DEF MAX_NDIM = 25
 
@@ -2190,45 +2194,31 @@ def array_split(ndarray ary, indices_or_sections, Py_ssize_t axis):
 cdef Py_ssize_t PY_SSIZE_T_MAX = sys.maxsize
 
 
-cdef class broadcast:
-    """Object that performs broadcasting.
-
-    CuPy actually uses this class to support broadcasting in various
-    operations. Note that this class does not provide an iterator.
-
-    Args:
-        arrays (tuple of arrays): Arrays to be broadcasted.
-
-    Attributes:
-        ~broadcast.shape (tuple of ints): The broadcasted shape.
-        nd (int): Number of dimensions of the broadcasted shape.
-        ~broadcast.size (int): Total size of the broadcasted shape.
-        values (list of arrays): The broadcasted arrays.
-
-    .. seealso:: :class:`numpy.broadcast`
-
-    """
+cdef class _broadcast_impl:
 
     cdef:
-        readonly tuple values
-        readonly tuple shape
+        vector.vector[Py_ssize_t] _shape
         readonly Py_ssize_t size
         readonly Py_ssize_t nd
+        readonly list strides_list
 
-    def __init__(self, *arrays):
+    def __init__(self, arg_infos):
         cdef Py_ssize_t i, j, s, smin, smax, a_ndim, a_sh
-        cdef vector.vector[Py_ssize_t] shape, strides, r_shape, r_strides
         cdef vector.vector[vector.vector[Py_ssize_t]] shape_arr
-        cdef ndarray a, view
+        cdef vector.vector[Py_ssize_t] strides, r_strides, r_shape, a_shape
+        cdef ArgInfo arg
+
         rev = slice(None, None, -1)
 
         self.nd = 0
-        for x in arrays:
-            if not isinstance(x, ndarray):
+        for arg in arg_infos:
+            if arg is None:
+
                 continue
-            a = x
-            self.nd = max(self.nd, <Py_ssize_t>a._shape.size())
-            r_shape.assign(a._shape.rbegin(), a._shape.rend())
+            if not arg.is_ndarray:
+                continue
+            self.nd = max(self.nd, arg.ndim)
+            r_shape = reversed(arg.shape)
             shape_arr.push_back(r_shape)
 
         r_shape.clear()
@@ -2246,35 +2236,118 @@ cdef class broadcast:
                     'single shape')
             r_shape.push_back(0 if smin == 0 else smax)
 
-        shape.assign(r_shape.rbegin(), r_shape.rend())
-        self.shape = tuple(shape)
-        self.size = internal.prod_ssize_t(shape)
+        self._shape.assign(r_shape.rbegin(), r_shape.rend())
+        self.size = internal.prod_ssize_t(self._shape)
 
-        broadcasted = []
-        for x in arrays:
-            if not isinstance(x, ndarray):
-                broadcasted.append(x)
+        # `None` means the original argument can be used without
+        # chainging strides.
+        strides_list = []
+
+        for arg in arg_infos:
+            if arg is None:
+                strides_list.append(None)
                 continue
-            a = x
-            if internal.vector_equal(a._shape, shape):
-                broadcasted.append(a)
+            if not arg.is_ndarray:
+                strides_list.append(None)
+                continue
+
+            a_shape = arg.shape
+            if internal.vector_equal(a_shape, self._shape):
+                strides_list.append(None)
                 continue
 
             r_strides.assign(self.nd, <Py_ssize_t>0)
-            a_ndim = a._shape.size()
+            a_ndim = arg.ndim
+            strides_ = arg.strides
             for i in range(a_ndim):
-                a_sh = a._shape[a_ndim - i - 1]
+                a_sh = a_shape[a_ndim - i - 1]
                 if a_sh == r_shape[i]:
-                    r_strides[i] = a._strides[a_ndim - i - 1]
+                    r_strides[i] = strides_[a_ndim - i - 1]
                 elif a_sh != 1:
                     raise ValueError('Broadcasting failed')
 
             strides.assign(r_strides.rbegin(), r_strides.rend())
-            view = a.view()
-            view._set_shape_and_strides(shape, strides)
-            broadcasted.append(view)
+            strides_list.append(strides)
 
-        self.values = tuple(broadcasted)
+        self.strides_list = strides_list
+
+    cpdef tuple shape(self):
+        return tuple(self._shape)
+
+    cpdef list apply(self, arrays):
+        cdef ndarray view
+        cdef vector.vector[Py_ssize_t] strides
+
+        assert len(arrays) == len(self.strides_list), \
+            (len(arrays), len(self.strides_list))
+
+        broadcasted = []
+        for i in range(<Py_ssize_t>len(arrays)):
+            x = arrays[i]
+            strides_ = self.strides_list[i]
+            if strides_ is None:
+                broadcasted.append(x)
+            else:
+                strides = <vector.vector[Py_ssize_t]>strides_
+                view = (<ndarray>x).view()
+                view._set_shape_and_strides(self._shape, strides)
+                broadcasted.append(view)
+
+        return broadcasted
+
+    cpdef list apply_infos(self, arg_infos):
+        cdef ArgInfo a
+
+        assert len(arg_infos) == len(self.strides_list)
+
+        broadcasted = []
+        for i in range(<Py_ssize_t>len(arg_infos)):
+            a = arg_infos[i]
+            strides_ = self.strides_list[i]
+            if strides_ is None:
+                broadcasted.append(a)
+            else:
+                strides = <vector.vector[Py_ssize_t]>strides_
+                shape = tuple(self._shape)
+                broadcasted.append(
+                    ArgInfo(None, ndarray, a.dtype, shape, len(shape),
+                            tuple(strides)))
+
+        return broadcasted
+
+
+cdef class broadcast:
+    """Object that performs broadcasting.
+
+    CuPy actually uses this class to support broadcasting in various
+    operations. Note that this class does not provide an iterator.
+
+    Args:
+        arrays (tuple of arrays): Arrays to be broadcasted.
+
+    Attributes:
+        shape (tuple of ints): The broadcasted shape.
+        nd (int): Number of dimensions of the broadcasted shape.
+        size (int): Total size of the broadcasted shape.
+        values (list of arrays): The broadcasted arrays.
+
+    .. seealso:: :class:`numpy.broadcast`
+
+    """
+
+    cdef:
+        readonly tuple values
+        readonly tuple shape
+        readonly Py_ssize_t size
+        readonly Py_ssize_t nd
+
+    def __init__(self, *arrays):
+        arg_infos = ArgInfo_from_args(arrays, True)
+        impl = _broadcast_impl(arg_infos)
+        self.values = tuple(impl.apply(arrays))
+        self.shape = tuple(impl.shape())
+        self.size = impl.size
+        self.nd = impl.nd
 
 
 cpdef ndarray broadcast_to(ndarray array, shape):
@@ -3454,7 +3527,6 @@ cpdef ndarray matmul(ndarray a, ndarray b, ndarray out=None):
         return ret
 
 
-cdef _cuda_runtime_version = None
 cdef _tensordot_core_mul_sum = ReductionKernel(
     'S x, T y', 'U out',
     'static_cast<U>(x) * static_cast<U>(y)',
@@ -3480,15 +3552,11 @@ cpdef ndarray tensordot_core(
         out.fill(0)
         return out
 
-    global _cuda_runtime_version
-    if _cuda_runtime_version is None:
-        _cuda_runtime_version = runtime.runtimeGetVersion()
-
-    use_sgemmEx = (_cuda_runtime_version >= 7500 and
+    use_sgemmEx = (cuda.get_runtime_version() >= 7500 and
                    a.dtype == 'e' and b.dtype == 'e' and
                    (ret_dtype == 'e' or ret_dtype == 'f'))
     use_tensor_core = (use_sgemmEx and
-                       _cuda_runtime_version >= 9000 and
+                       cuda.get_runtime_version() >= 9000 and
                        int(device.get_compute_capability()) == 70)
 
     if use_sgemmEx or ret_dtype in 'fdFD':
@@ -4032,7 +4100,7 @@ def _inclusive_scan_kernel(dtype, block_size):
     """
 
     name = "inclusive_scan_kernel"
-    dtype = _get_typename(dtype)
+    dtype = _get_ctype_name(dtype)
     source = string.Template("""
     extern "C" __global__ void ${name}(const CArray<${dtype}, 1> src,
         CArray<${dtype}, 1> dst){
@@ -4072,14 +4140,14 @@ def _inclusive_scan_kernel(dtype, block_size):
         }
     }
     """).substitute(name=name, dtype=dtype, block_size=block_size)
-    module = compile_with_cache(source)
+    module = compile.compile_with_cache(source)
     return module.get_function(name)
 
 
 @util.memoize(for_each_device=True)
 def _add_scan_blocked_sum_kernel(dtype):
     name = "add_scan_blocked_sum_kernel"
-    dtype = _get_typename(dtype)
+    dtype = _get_ctype_name(dtype)
     source = string.Template("""
     extern "C" __global__ void ${name}(CArray<${dtype}, 1> src_dst){
         long long n = src_dst.size();
@@ -4092,15 +4160,15 @@ def _add_scan_blocked_sum_kernel(dtype):
         }
     }
     """).substitute(name=name, dtype=dtype)
-    module = compile_with_cache(source)
+    module = compile.compile_with_cache(source)
     return module.get_function(name)
 
 
 @util.memoize(for_each_device=True)
 def _nonzero_1d_kernel(src_dtype, index_dtype):
     name = "nonzero_1d_kernel"
-    src_dtype = _get_typename(src_dtype)
-    index_dtype = _get_typename(index_dtype)
+    src_dtype = _get_ctype_name(src_dtype)
+    index_dtype = _get_ctype_name(index_dtype)
 
     source = string.Template("""
     extern "C" __global__ void ${name}(const CArray<${src_dtype}, 1> src,
@@ -4115,16 +4183,16 @@ def _nonzero_1d_kernel(src_dtype, index_dtype):
         }
     }
     """).substitute(name=name, src_dtype=src_dtype, index_dtype=index_dtype)
-    module = compile_with_cache(source)
+    module = compile.compile_with_cache(source)
     return module.get_function(name)
 
 
 @util.memoize(for_each_device=True)
 def _nonzero_kernel(src_dtype, src_ndim, index_dtype, dst_dtype):
     name = "nonzero_kernel"
-    src_dtype = _get_typename(src_dtype)
-    index_dtype = _get_typename(index_dtype)
-    dst_dtype = _get_typename(dst_dtype)
+    src_dtype = _get_ctype_name(src_dtype)
+    index_dtype = _get_ctype_name(index_dtype)
+    dst_dtype = _get_ctype_name(dst_dtype)
 
     source = string.Template("""
         extern "C" __global__ void ${name}(const CArray<${src_dtype}, 1> src,
@@ -4150,7 +4218,7 @@ def _nonzero_kernel(src_dtype, src_ndim, index_dtype, dst_dtype):
         """).substitute(name=name, src_dtype=src_dtype,
                         src_ndim=src_ndim, index_dtype=index_dtype,
                         dst_dtype=dst_dtype)
-    module = compile_with_cache(source)
+    module = compile.compile_with_cache(source)
     return module.get_function(name)