--- a/cupy/core/fusion.py
+++ b/cupy/core/fusion.py
@@ -551,7 +551,7 @@ def _get_fusion(func, nin, reduce, post_map, identity, input_types, name):
         submodules = _gather_submodules(op_list + post_ops)
         submodule_code = ''.join(_get_submodule_code(v)
                                  for v in submodules.values())
-        submodule_code += reduce._raw._preamble + pre_code + post_code
+        submodule_code += reduce._raw.preamble + pre_code + post_code
         operation_args = ['v' + str(i) for i in six.moves.range(nin)]
         operation = '_pre_map(' + ', '.join(operation_args) + ')'
         out_params = '%s res' % post_out.ty
@@ -677,11 +677,7 @@ class ufunc(core.ufunc):
         self.nin = fusion_op.nin
         self.nout = fusion_op.nout
         self.nargs = fusion_op.nargs
-        self._ops = fusion_op._ops
-        self._preamble = fusion_op._preamble
         self.__doc__ = cupy_op.__doc__
-        self._params = fusion_op._params
-        self._routine_cache = fusion_op._routine_cache
 
         self._fusion_op = fusion_op
         self._cupy_op = cupy_op