… cupy/core/reduction.pxi
… cupy/cuda/__init__.py