Refactor kernel code generation and invocation.

### Keys:
1. The code for launching kernels were duplicated (i.e. between `ElementwiseKernel` and `ufunc`, and between `ReductionKernel` and `simple_reduction_function`). I factored out common logic and merged the diverged code.
2. Encapsulated argument metadata (as `ArgInfo`). This is aimed at fusion code refactoring.
    * Divided `broadcast` to the metadata retrieval (`_broadcast_impl.__init__`) and application to actual arrays (`_broadcast_impl.apply`).
3. Refactored kernel code emission (with `KernelCodeEmitter`) for elementary operations (both elementwise and reduction) in modularized manner. This is also aimed at fusion.

### Current status:
Currently performance is degraded. That's why this PR is WIP.

Left column:  blocking time of Python code
Right column: GPU time measured by `cupy.cuda.get_elapsed_time`
(Measured with https://github.com/niboshi/cupy-perf)

```
Original [master @ 630ace19bfec4f61c9518401cbd8d577bdf5e765]
empty               :     0.200 us   +/- 0.014 us      5.343 us   +/- 0.410 us
sum                 :    35.549 us   +/- 0.987 us     42.337 us   +/- 1.176 us
sum_huge            :    42.373 us   +/- 9.522 us  10484.689 us   +/-14.838 us
add                 :    25.316 us   +/- 1.726 us     31.822 us   +/- 2.067 us
add_out             :    17.830 us   +/- 1.543 us     24.636 us   +/- 2.056 us
add_out_huge        :    21.851 us   +/- 4.426 us   9377.292 us   +/- 4.485 us
userkernel_create   :     2.353 us   +/- 0.625 us      7.757 us   +/- 1.014 us
userkenel           :    12.217 us   +/- 0.593 us     19.725 us   +/- 1.522 us
userkernel_huge     :    15.893 us   +/- 4.136 us   9371.349 us   +/- 4.243 us

```
```
This PR
empty               :     0.181 us   +/- 0.009 us      5.371 us   +/- 0.322 us
sum                 :    41.038 us   +/- 2.466 us     48.072 us   +/- 2.604 us
sum_huge            :    46.439 us   +/- 9.223 us  10484.955 us   +/-12.350 us
add                 :    29.894 us   +/- 0.631 us     36.599 us   +/- 0.706 us
add_out             :    21.656 us   +/- 0.450 us     28.692 us   +/- 0.596 us
add_out_huge        :    26.021 us   +/- 7.232 us   9381.202 us   +/- 7.555 us
userkernel_create   :     1.038 us   +/- 0.171 us      6.407 us   +/- 0.638 us
userkenel           :    15.110 us   +/- 1.232 us     22.115 us   +/- 1.913 us
userkernel_huge     :    18.648 us   +/- 5.055 us   9373.585 us   +/- 5.171 us
```
\* `empty` means empty Python function calls