@okuta 2017-10-29
IMO, this performance degradation can be accepted.

Please proceed with your work.