@okuta 2018-01-08
I want to merge this PR.
If you have no time, I can improve it.