@grafi-tt 2018-04-09
@niboshi I'd like to see a functionality to perform multiple reduction within one ReductionKernel. One notable use case is to compute L1 and L2 moments simultaneously. Currently, two invocation of ReductionKernel is required to compute variance and mean of data. (Example:  https://github.com/chainer/chainer/blob/v3.5.0/chainer/functions/normalization/layer_normalization.py#L39, where `mean` is called twice.) But it causes unnecessary memory accesses. Performance loss is not negligible, especially for ResNet-like CNNs where convolutions are relatively small.

Are you intended to implement this functionality in this PR?  If not, should I open an issue for the functionality?

Many thanks.