@niboshi 2020-02-13
Closing as improvements in this PR either has been already merged or being discussed in other PRs.
(e.g. #2801  #2920  #2940)