@niboshi 2020-02-13
@grafi-tt 
I'm sorry for a long absence in this PR...
This PR was only meant for refactoring, and no feature addition were intended.